Extract UpdateToInsert
